api re-integration ATD - quran, code little organized - directory structure etc
