+++ components/HomeScreens/Quran/QuranTracker.js
@@ -10,17 +10,31 @@ import {colors} from '../../../assets/colors/colors';
 import {FontSize} from '../../../assets/fonts/fonts';
 // components
 import QuranTrackerView from './QuranTrackerView';
+import {getArabicDate} from '../../../redux-toolkit/features/arabic-date/arabicDate';
+import {useSelector} from 'react-redux';
 
 const QuranTracker = () => {
+  const day = useSelector(getArabicDate);
+
   // todo:perf: memoize other components, so that useState doesn't affect them all
-  const {data = {}, error, isError, isLoading} = useGetRecitationInfoQuery();
+  const {
+    data = {},
+    error,
+    isError,
+    isLoading,
+  } = useGetRecitationInfoQuery({
+    year: parseInt(day.year, 10),
+    month: parseInt(day.monthNumber, 10),
+    day: parseInt(day.day, 10),
+  });
+
   useEffect(() => {
     try {
-      if (isError) {
-        console.error('SCREEN:QURAN: get recitation error: ', error);
-      }
+      if (!isLoading && !(JSON.stringify(data) === '{}')) {
+        if (isError) {
+          console.error('SCREEN:QURAN: get recitation error: ', error);
+        }
 
-      if (!isLoading) {
         // console.log('SCREEN:QURAN: get recitation data: ', data);
       }
     } catch (issue) {
